+++ R/imputation.R
@@ -161,6 +161,10 @@
 ##' Data Sets to Compare Imputation Strategies*. J Proteome Res. 2016 Apr
 ##' 1;15(4):1116-25. doi: 10.1021/acs.jproteome.5b00981. PubMed PMID:26906401.
 ##'
+##' @seealso The *Imputation* vignette in the `QFeatures` package available with
+##'     `vignette("Imputation", package = "QFeatures")` and online at
+##'     <https://rformassspectrometry.github.io/QFeatures/articles/Imputation.html>.
+##'
 ##' @rdname imputation
 ##'
 ##' @aliases imputeMethods impute_neighbour_average impute_knn impute_mle impute_bpca impute_mixed impute_min impute_zero impute_with impute_matrix impute_MinDet impute_MinProb impute_QRILC
@@ -403,16 +407,30 @@ impute_RF <- function(x, MARGIN = 2L, ...) {
 ##'     random. Only relevant when `methods` is `mixed`.
 ##'
 ##' @param mar Imputation method for values missing at random. See `method`
-##'     above.
+##'     above. Only relevant when `methods` is `mixed`.
 ##'
 ##' @param mnar Imputation method for values missing not at random. See `method`
-##'     above.
+##'     above. Only relevant when `methods` is `mixed`.
+##'
+##' @param marArgs `list()` of arguments to be passed to the `mar` imputation
+##'     function. Only relevant when `methods` is `mixed`.
+##'
+##' @param mnarArgs `list()` of arguments to be passed to the `mnar` imputation
+##'     function. Only relevant when `methods` is `mixed`.
+##'
+##' @param split `logical(1)` defining if the MAR and MNAR sub-matrices are to
+##'     be split before imputation (default is `TRUE`), or if the whole data
+##'     should be used to compute imputed values for the individual
+##'     sub-matrices. Only relevant when `methods` is `mixed`.
 ##'
 ##' @export
 ##'
 ##' @rdname imputation
 impute_mixed <- function(x, randna, mar, mnar,
-                         MARGIN = c(1L, 1L)) {
+                         MARGIN = c(1L, 1L),
+                         marArgs = list(),
+                         mnarArgs = list(),
+                         split = TRUE) {
     if (missing(randna))
         stop("Mixed imputation requires 'randna' argument. See ?impute_mixed.",
              call. = FALSE)
@@ -427,16 +445,39 @@ impute_mixed <- function(x, randna, mar, mnar,
     if (length(MARGIN) == 1)
         MARGIN <- c(MARGIN, MARGIN)
     if (length(MARGIN) != 2)
-        stop("MARGIN must be of length 1 or 2.")
+        stop("MARGIN must be of length 1 or 2.", call. = FALSE)
     if (length(randna) != nrow(x))
         stop("Number of rows and length of randna must be equal.",
              call. = FALSE)
-    ## MAR imputation - first MARGIN
-    x[randna, ] <- impute_matrix(x[randna, ], mar,
-                                 MARGIN = MARGIN[[1]])
-    ## MAR imputation - second MARGIN
-    x[!randna, ] <- impute_matrix(x[!randna, ], mnar,
-                                  MARGIN = MARGIN[[2]])
+    if (split) { ## impute separately
+        ## MAR imputation - first MARGIN
+        args <- append(list(x = x[randna, ],
+                            method = mar,
+                            MARGIN = MARGIN[[1]]),
+                       marArgs)
+        x[randna, ] <- do.call(impute_matrix, args)
+        ## MNAR imputation - second MARGIN
+        args <- append(list(x = x[!randna, ],
+                            method = mnar,
+                            MARGIN = MARGIN[[2]]),
+                       mnarArgs)
+        x[!randna, ] <- do.call(impute_matrix, args)
+    } else { ## use the whole matrix for imputation
+        ## MAR imputation - first MARGIN
+        args <- append(list(x = x,
+                            method = mar,
+                            MARGIN = MARGIN[[1]]),
+                       marArgs)
+        ximp <- do.call(impute_matrix, args)
+        x[randna, ] <- ximp[randna, ]
+        ## MNAR imputation - second MARGIN
+        args <- append(list(x = x,
+                            method = mnar,
+                            MARGIN = MARGIN[[2]]),
+                       mnarArgs)
+        ximp <- do.call(impute_matrix, args)
+        x[!randna, ] <- ximp[!randna, ]
+    }
     x
 }
 
@@ -570,11 +611,16 @@ impute_fun <- function(x, FUN, MARGIN = 1L, ...) {
 ##' getImputeMargin(impute_zero) ## NA: no margin here
 ##'
 ##' ## default margin for all MsCoreUtils::impute_* functions
-##' sapply(ls("package:MsCoreUtils", pattern = "impute_"), getImputeMargin)
+##' getImputeMargin()
 getImputeMargin <- function(fun) {
+    if (missing(fun))
+        return(.getAllImputeMargin())
     args <- formals(fun)
     i <- grep("MARGIN", names(args))
     if (length(i)) ans <- args[[i]]
     else ans <- NA
     ans
 }
+
+.getAllImputeMargin <- function()
+    sapply(ls("package:MsCoreUtils", pattern = "impute_"), getImputeMargin)